Minify bundles in prod builds
